Add implementation of allEvenBits()
The following is explanation of this implementation. After copied
the arg to local var, we then bitwise AND local var with a
constant, which is to make sure that the remained bit is all
even-numbered bits, after bitwise AND, we then perform a bitwise
XOR with another constant, here we get 0 if all even-numbered bits
is set to 1, vice versa, then we logical invert the result to
the required output then return it.